Added an inventory system!
- Added an inventory that adds and removes items, as well as checks that
they are there. It works in distinct item slots which have customizable
sizes, and there is a customizable number of slots available.
- The inventory can also be sorted in order of increasing item ID and
decreasing amount of items.
- So far I added Money, Health, Apples, and Oranges, but you can only
see the contents of the inventory by printing to the console as of right
now.
- Spiced up Lava Man quest to include giving apples as a reward, as well
as Squirtle taking 5 apples from you every time you have some. What a
jerk!